requirement.txt
Flask
gunicorn
google-generativeai
google-cloud-storage